+++ packages/core/client/src/schema-component/antd/gantt/components/gantt/gantt.tsx
@@ -133,7 +133,6 @@ export const Gantt: any = (props: any) => {
   useEffect(() => {
     tableCtx.field.onExpandClick = handleTableExpanderClick;
     tableCtx.field.onRowSelect = handleRowSelect;
-    tableCtx.setExpandFlag(true);
   }, []);
   useEffect(() => {
     expandAndCollapseAll?.(!expandFlag);
