+++ packages/core/client/src/schema-component/antd/table-v2/utils.ts
@@ -17,16 +17,3 @@ export function extractIndex(str) {
   });
   return numbers.join('.');
 }
-
-export function getIdsWithChildren(nodes) {
-  const ids = [];
-  if (nodes) {
-    for (const node of nodes) {
-      if (node?.children && node.children.length > 0) {
-        ids.push(node.id);
-        ids.push(...getIdsWithChildren(node?.children));
-      }
-    }
-  }
-  return ids;
-}
